+++ wrappersnap/wasdishared/src/main/java/wasdi/shared/utils/Utils.java
@@ -142,15 +142,22 @@ public class Utils {
 		return sReturn;
 	}
 
-	public static String GetFileNameExtension(String sInputFile) {
+	/**
+	 * Extracts the extension fro a file name
+	 * @param sInputFile Input File Name
+	 * @return Extension or empty string
+	 */
+	public static String getFileNameExtension(String sInputFile) {
 		String sReturn = "";
 		File oFile = new File(sInputFile);
 		String sInputFileNameOnly = oFile.getName();
 
-		// Create a clean layer id: the file name without any extension
-		String[] asLayerIdSplit = sInputFileNameOnly.split("\\.");
-		if (asLayerIdSplit != null && asLayerIdSplit.length > 0) {
-			sReturn = asLayerIdSplit[asLayerIdSplit.length - 1];
+		if (sInputFileNameOnly.contains(".")) {
+			// Create a clean layer id: the file name without any extension
+			String[] asLayerIdSplit = sInputFileNameOnly.split("\\.");
+			if (asLayerIdSplit != null && asLayerIdSplit.length > 0) {
+				sReturn = asLayerIdSplit[asLayerIdSplit.length - 1];
+			}			
 		}
 
 		return sReturn;